Move volume to bin/glusterd2/volume
