+++ bin/glusterd2/commands/volumes/volume-create_test.go
@@ -8,11 +8,11 @@ import (
 
 	"github.com/gluster/glusterd2/bin/glusterd2/peer"
 	"github.com/gluster/glusterd2/bin/glusterd2/transaction"
+	"github.com/gluster/glusterd2/bin/glusterd2/volume"
 	"github.com/gluster/glusterd2/brick"
 	"github.com/gluster/glusterd2/pkg/api"
 	gderrors "github.com/gluster/glusterd2/pkg/errors"
 	"github.com/gluster/glusterd2/tests"
-	"github.com/gluster/glusterd2/volume"
 
 	"github.com/pborman/uuid"
