+++ bin/glusterd2/commands/volumes/volume-delete.go
@@ -7,7 +7,7 @@ import (
 	restutils "github.com/gluster/glusterd2/bin/glusterd2/servers/rest/utils"
 	"github.com/gluster/glusterd2/bin/glusterd2/transaction"
 	"github.com/gluster/glusterd2/bin/glusterd2/volgen"
-	"github.com/gluster/glusterd2/volume"
+	"github.com/gluster/glusterd2/bin/glusterd2/volume"
 
 	"github.com/gorilla/mux"
 	"github.com/pborman/uuid"
