+++ bin/glusterd2/commands/volumes/volume-info.go
@@ -4,8 +4,8 @@ import (
 	"net/http"
 
 	restutils "github.com/gluster/glusterd2/bin/glusterd2/servers/rest/utils"
+	"github.com/gluster/glusterd2/bin/glusterd2/volume"
 	"github.com/gluster/glusterd2/pkg/errors"
-	"github.com/gluster/glusterd2/volume"
 
 	"github.com/gorilla/mux"
 )
