+++ bin/glusterd2/commands/volumes/volume-list.go
@@ -4,7 +4,7 @@ import (
 	"net/http"
 
 	restutils "github.com/gluster/glusterd2/bin/glusterd2/servers/rest/utils"
-	"github.com/gluster/glusterd2/volume"
+	"github.com/gluster/glusterd2/bin/glusterd2/volume"
 )
 
 func volumeListHandler(w http.ResponseWriter, r *http.Request) {
