+++ bin/glusterd2/commands/volumes/volume-start.go
@@ -6,8 +6,8 @@ import (
 	"github.com/gluster/glusterd2/bin/glusterd2/gdctx"
 	restutils "github.com/gluster/glusterd2/bin/glusterd2/servers/rest/utils"
 	"github.com/gluster/glusterd2/bin/glusterd2/transaction"
+	"github.com/gluster/glusterd2/bin/glusterd2/volume"
 	"github.com/gluster/glusterd2/pkg/errors"
-	"github.com/gluster/glusterd2/volume"
 
 	"github.com/gorilla/mux"
 	"github.com/pborman/uuid"
