+++ bin/glusterd2/volgen/graph.go
@@ -6,8 +6,8 @@ import (
 	"path"
 	"strings"
 
+	"github.com/gluster/glusterd2/bin/glusterd2/volume"
 	"github.com/gluster/glusterd2/pkg/utils"
-	"github.com/gluster/glusterd2/volume"
 	"github.com/gluster/glusterd2/xlator"
 )
