+++ bin/glusterd2/volgen/volgen.go
@@ -9,9 +9,9 @@ import (
 	"strings"
 
 	"github.com/gluster/glusterd2/bin/glusterd2/store"
+	"github.com/gluster/glusterd2/bin/glusterd2/volume"
 	"github.com/gluster/glusterd2/brick"
 	"github.com/gluster/glusterd2/pkg/utils"
-	"github.com/gluster/glusterd2/volume"
 )
 
 const (
